Program is complete
Removed debugging statements.
When ran, the program will connect to the server at "codebank.xyz" with
port 38001.
Next, it will prompt user to enter a username; if username is taken,
server will respond with "Name in use." and program will quit.
Otherwise, program will infinitely allow user to enter messages to send
to all clients connected to the server.